Move the Decodable trait to the consensus module
Now we have a `decode` module move the `consensus::Decodable` trait to
it. Add public re-export and don't change any other code, to assist
review.
