+++ bitcoin/src/consensus/encode.rs
@@ -23,7 +23,7 @@ use hashes::{sha256, sha256d, Hash};
 use crate::bip152::{PrefilledTransaction, ShortId};
 use crate::blockdata::transaction::{Transaction, TxIn, TxOut};
 // TODO(Tobin): Remove this re-export once consensus refactor is done.
-pub use crate::consensus::decode::Error;
+pub use crate::consensus::decode::{Decodable, Error, MAX_VEC_SIZE};
 use crate::hash_types::{BlockHash, FilterHash, FilterHeader, TxMerkleNode};
 use crate::io::{self, Cursor, Read};
 #[cfg(feature = "std")]
@@ -190,9 +190,6 @@ impl<R: Read + ?Sized> ReadExt for R {
     }
 }
 
-/// Maximum size, in bytes, of a vector we are allowed to decode.
-pub const MAX_VEC_SIZE: usize = 4_000_000;
-
 /// Data which can be encoded in a consensus-consistent way.
 pub trait Encodable {
     /// Encodes an object with a well-defined format.
@@ -202,60 +199,6 @@ pub trait Encodable {
     /// The number of bytes written on success. The only errors returned are errors propagated from
     /// the writer.
     fn consensus_encode<W: io::Write + ?Sized>(&self, writer: &mut W) -> Result<usize, io::Error>;
-}
-
-/// Data which can be encoded in a consensus-consistent way.
-pub trait Decodable: Sized {
-    /// Decode `Self` from a size-limited reader.
-    ///
-    /// Like `consensus_decode` but relies on the reader being limited in the amount of data it
-    /// returns, e.g. by being wrapped in [`std::io::Take`].
-    ///
-    /// Failing to abide to this requirement might lead to memory exhaustion caused by malicious
-    /// inputs.
-    ///
-    /// Users should default to `consensus_decode`, but when data to be decoded is already in a byte
-    /// vector of a limited size, calling this function directly might be marginally faster (due to
-    /// avoiding extra checks).
-    ///
-    /// ### Rules for trait implementations
-    ///
-    /// * Simple types that that have a fixed size (own and member fields), don't have to overwrite
-    ///   this method, or be concern with it.
-    /// * Types that deserialize using externally provided length should implement it:
-    ///   * Make `consensus_decode` forward to `consensus_decode_bytes_from_finite_reader` with the
-    ///     reader wrapped by `Take`. Failure to do so, without other forms of memory exhaustion
-    ///     protection might lead to resource exhaustion vulnerability.
-    ///   * Put a max cap on things like `Vec::with_capacity` to avoid oversized allocations, and
-    ///     rely on the reader running out of data, and collections reallocating on a legitimately
-    ///     oversized input data, instead of trying to enforce arbitrary length limits.
-    /// * Types that contain other types that implement custom
-    ///   `consensus_decode_from_finite_reader`, should also implement it applying same rules, and
-    ///   in addition make sure to call `consensus_decode_from_finite_reader` on all members, to
-    ///   avoid creating redundant `Take` wrappers. Failure to do so might result only in a tiny
-    ///   performance hit.
-    #[inline]
-    fn consensus_decode_from_finite_reader<R: io::Read + ?Sized>(
-        reader: &mut R,
-    ) -> Result<Self, Error> {
-        // This method is always strictly less general than, `consensus_decode`, so it's safe and
-        // make sense to default to just calling it. This way most types, that don't care about
-        // protecting against resource exhaustion due to malicious input, can just ignore it.
-        Self::consensus_decode(reader)
-    }
-
-    /// Decode an object with a well-defined format.
-    ///
-    /// This is the method that should be implemented for a typical, fixed sized type
-    /// implementing this trait. Default implementation is wrapping the reader
-    /// in [`crate::io::Take`] to limit the input size to [`MAX_VEC_SIZE`], and forwards the call to
-    /// [`Self::consensus_decode_from_finite_reader`], which is convenient
-    /// for types that override [`Self::consensus_decode_from_finite_reader`]
-    /// instead.
-    #[inline]
-    fn consensus_decode<R: io::Read + ?Sized>(reader: &mut R) -> Result<Self, Error> {
-        Self::consensus_decode_from_finite_reader(reader.take(MAX_VEC_SIZE as u64).by_ref())
-    }
 }
 
 /// A variable-length unsigned integer.
